Second stage refactoring, normalizing initialization patterns and renaming tests to match implementation classes they serve.
