+++ geomesa-spark/geomesa-spark-jts/src/main/scala/org/locationtech/geomesa/spark/jts/encoders/SpatialEncoders.scala
@@ -23,7 +23,3 @@ trait SpatialEncoders {
   implicit def jtsGeometryCollectionEncoder: Encoder[GeometryCollection] = ExpressionEncoder()
 }
 
-object SpatialEncoders extends SpatialEncoders  {
-
-}
-
