+++ geomesa-spark/geomesa-spark-jts/src/main/scala/org/locationtech/geomesa/spark/jts/rules/GeometryLiteralRules.scala
@@ -8,33 +8,15 @@
 
 package org.locationtech.geomesa.spark.jts.rules
 
-import com.vividsolutions.jts.geom.Geometry
 import org.apache.spark.sql.SQLContext
-import org.apache.spark.sql.catalyst.InternalRow
-import org.apache.spark.sql.catalyst.expressions.codegen.CodegenFallback
-import org.apache.spark.sql.catalyst.expressions.{GenericInternalRow, LeafExpression, Literal, ScalaUDF}
+import org.apache.spark.sql.catalyst.expressions.{GenericInternalRow, Literal, ScalaUDF}
 import org.apache.spark.sql.catalyst.plans.logical._
 import org.apache.spark.sql.catalyst.rules.Rule
 import org.apache.spark.sql.jts.GeometryUDT
-import org.apache.spark.sql.jts.JTSTypes._
-import org.apache.spark.sql.types.DataType
 
 import scala.util.Try
 
-object JTSRules {
-
-  // new AST expressions
-  case class GeometryLiteral(repr: InternalRow, geom: Geometry) extends LeafExpression  with CodegenFallback {
-
-    override def foldable: Boolean = true
-
-    override def nullable: Boolean = true
-
-    override def eval(input: InternalRow): Any = repr
-
-    override def dataType: DataType = GeometryTypeInstance
-  }
-
+object GeometryLiteralRules {
 
   object ScalaUDFRule extends Rule[LogicalPlan] {
     override def apply(plan: LogicalPlan): LogicalPlan = {
@@ -56,7 +38,7 @@ object JTSRules {
     }
   }
 
-  def registerOptimizations(sqlContext: SQLContext): Unit = {
+  private[jts] def registerOptimizations(sqlContext: SQLContext): Unit = {
     Seq(ScalaUDFRule).foreach { r =>
       if(!sqlContext.experimental.extraOptimizations.contains(r))
         sqlContext.experimental.extraOptimizations ++= Seq(r)
