+++ geomesa-spark/geomesa-spark-jts/src/main/scala/org/locationtech/geomesa/spark/jts/rules/package.scala
@@ -1,4 +1,3 @@
-
 /***********************************************************************
  * Copyright (c) 2013-2018 Commonwealth Computer Research, Inc.
  * All rights reserved. This program and the accompanying materials
@@ -7,20 +6,12 @@
  * http://www.opensource.org/licenses/apache2.0.php.
  ***********************************************************************/
 
-package org.locationtech.geomesa.spark.util
+package org.locationtech.geomesa.spark.jts
 
-import org.apache.commons.pool.impl.GenericObjectPool
+import org.apache.spark.sql.SQLContext
 
-// NB: This is a duplicate of a file in geomesa-utils
-trait ObjectPoolUtils[A] {
-  val pool: GenericObjectPool[A]
-
-  def withResource[B](f: A => B): B = {
-    val obj = pool.borrowObject()
-    try {
-      f(obj)
-    } finally {
-      pool.returnObject(obj)
-    }
+package object rules {
+  def registerOptimizations(sqlContext: SQLContext): Unit = {
+    GeometryLiteralRules.registerOptimizations(sqlContext)
   }
 }
