+++ geomesa-spark/geomesa-spark-jts/src/main/scala/org/locationtech/geomesa/spark/jts/udf/GeometricOutputFunctions.scala
@@ -10,14 +10,12 @@
 package org.locationtech.geomesa.spark.jts.udf
 
 import com.vividsolutions.jts.geom.{Geometry, Point}
-import org.apache.spark.sql.{Column, SQLContext, TypedColumn}
+import org.apache.spark.sql.SQLContext
 import org.geotools.geojson.geom.GeometryJSON
-import org.locationtech.geomesa.spark.util.SQLFunctionHelper._
-import org.locationtech.geomesa.spark.util.WKBUtils
-import org.locationtech.geomesa.spark.jts.encoders.SpatialEncoders._
-import org.locationtech.geomesa.spark.jts.encoders.SparkDefaultEncoders._
+import org.locationtech.geomesa.spark.jts.util.SQLFunctionHelper._
+import org.locationtech.geomesa.spark.jts.util.WKBUtils
 
-object SQLGeometricOutputFunctions {
+object GeometricOutputFunctions {
   // use ThreadLocal to ensure thread safety
   private val geomJSON = new ThreadLocal[GeometryJSON]() {
     override def initialValue() = new GeometryJSON()
@@ -27,38 +25,18 @@ object SQLGeometricOutputFunctions {
   val ST_AsLatLonText: Point => String = nullableUDF(point => toLatLonString(point))
   val ST_AsText: Geometry => String = nullableUDF(geom => geom.toText)
 
-  private[geomesa] val namer = Map(
+  private[geomesa] val outputNames = Map(
     ST_AsBinary -> "st_asBinary",
     ST_AsGeoJSON -> "st_asGeoJSON",
     ST_AsLatLonText -> "st_asLatLonText",
     ST_AsText -> "st_asText"
   )
 
-  def st_asBinary(geom: Column): TypedColumn[Any, Array[Byte]] =
-    udfToColumn(ST_AsBinary, namer, geom)
-  def st_asBinary(geom: Geometry): TypedColumn[Any, Array[Byte]] =
-    udfToColumnLiterals(ST_AsBinary, namer, geom)
-
-  def st_asGeoJSON(geom: Column): TypedColumn[Any, String] =
-    udfToColumn(ST_AsGeoJSON, namer, geom)
-  def st_asGeoJSON(geom: Geometry): TypedColumn[Any, String] =
-    udfToColumnLiterals(ST_AsGeoJSON, namer, geom)
-
-  def st_asLatLonText(point: Column): TypedColumn[Any, String] =
-    udfToColumn(ST_AsLatLonText, namer, point)
-  def st_asLatLonText(point: Point): TypedColumn[Any, String] =
-    udfToColumnLiterals(ST_AsLatLonText, namer, point)
-
-  def st_asText(geom: Column): TypedColumn[Any, String] =
-    udfToColumn(ST_AsText, namer, geom)
-  def st_asText(geom: Geometry): TypedColumn[Any, String] =
-    udfToColumnLiterals(ST_AsText, namer, geom).as[String]
-
-  def registerFunctions(sqlContext: SQLContext): Unit = {
-    sqlContext.udf.register(namer(ST_AsBinary), ST_AsBinary)
-    sqlContext.udf.register(namer(ST_AsGeoJSON), ST_AsGeoJSON)
-    sqlContext.udf.register(namer(ST_AsLatLonText), ST_AsLatLonText)
-    sqlContext.udf.register(namer(ST_AsText), ST_AsText)
+  private[jts] def registerFunctions(sqlContext: SQLContext): Unit = {
+    sqlContext.udf.register(outputNames(ST_AsBinary), ST_AsBinary)
+    sqlContext.udf.register(outputNames(ST_AsGeoJSON), ST_AsGeoJSON)
+    sqlContext.udf.register(outputNames(ST_AsLatLonText), ST_AsLatLonText)
+    sqlContext.udf.register(outputNames(ST_AsText), ST_AsText)
   }
 
   private def toLatLonString(point: Point): String = {
