+++ geomesa-spark/geomesa-spark-jts/src/test/scala/org/locationtech/geomesa/spark/jts/GeometricCastFunctionsTest.scala
@@ -9,32 +9,21 @@
 package org.locationtech.geomesa.spark.jts
 
 import com.vividsolutions.jts.geom._
-import org.apache.spark.sql._
 import org.apache.spark.sql.functions.lit
-import org.apache.spark.sql.jts.JTSTypes
 import org.junit.runner.RunWith
-import org.locationtech.geomesa.spark.jts.udf.SQLGeometricCastFunctions._
-import org.locationtech.geomesa.spark.jts.udf.SQLGeometricConstructorFunctions._
 import org.specs2.mutable.Specification
 import org.specs2.runner.JUnitRunner
 
 @RunWith(classOf[JUnitRunner])
-class SparkSQLGeometricCastTest extends Specification with BlankDataFrame {
+class GeometricCastFunctionsTest extends Specification with TestEnvironment {
 
   "sql geometry accessors" should {
     sequential
 
-    implicit var spark: SparkSession = null
-    var sc: SQLContext = null
-
     // before
     step {
-      spark = SparkSession.builder()
-        .appName("testSpark")
-        .master("local[*]")
-        .getOrCreate()
-      sc = spark.sqlContext
-      JTSTypes.init(sc)
+      // Trigger initialization of spark session
+      val _ = spark
     }
 
     "st_castToPoint" >> {
