+++ geomesa-spark/geomesa-spark-jts/src/test/scala/org/locationtech/geomesa/spark/jts/GeometricConstructorFunctionsTest.scala
@@ -9,34 +9,23 @@
 package org.locationtech.geomesa.spark.jts
 
 import com.vividsolutions.jts.geom._
-import org.apache.spark.sql._
 import org.apache.spark.sql.functions._
-import org.apache.spark.sql.jts.JTSTypes
 import org.geotools.geometry.jts.JTS
 import org.junit.runner.RunWith
-import org.locationtech.geomesa.spark.jts.udf.SQLGeometricCastFunctions._
-import org.locationtech.geomesa.spark.jts.udf.SQLGeometricConstructorFunctions._
-import org.locationtech.geomesa.spark.util.WKTUtils
+import org.locationtech.geomesa.spark.jts.util.WKTUtils
 import org.specs2.mutable.Specification
 import org.specs2.runner.JUnitRunner
 
 @RunWith(classOf[JUnitRunner])
-class SparkSQLGeometricConstructorsTest extends Specification with BlankDataFrame {
+class GeometricConstructorFunctionsTest extends Specification with TestEnvironment {
 
   "sql geometry constructors" should {
     sequential
 
-    implicit var spark: SparkSession = null
-    var sc: SQLContext = null
-
     // before
     step {
-      spark = SparkSession.builder()
-        .appName("testSpark")
-        .master("local[*]")
-        .getOrCreate()
-      sc = spark.sqlContext
-      JTSTypes.init(sc)
+      // Trigger initialization of spark session
+      val _ = spark
     }
 
     "st_geomFromWKT" >> {
@@ -55,7 +44,6 @@ class SparkSQLGeometricConstructorsTest extends Specification with BlankDataFram
       r.collect().head.getAs[Geometry](0) mustEqual expected
 
       dfBlank.select(st_geomFromWKT(point)).first mustEqual expected
-
     }
 
     "st_geometryFromText" >> {
