+++ geomesa-spark/geomesa-spark-sql/src/main/scala/org/apache/spark/sql/SQLGeometricConstructorFunctions.scala
@@ -11,7 +11,7 @@ package org.apache.spark.sql
 import com.vividsolutions.jts.geom._
 import org.geotools.geometry.jts.JTS
 import org.locationtech.geomesa.utils.geohash.{BoundingBox, GeoHash}
-import org.locationtech.geomesa.spark.util.SQLFunctionHelper.nullableUDF
+import org.locationtech.geomesa.spark.jts.util.SQLFunctionHelper.nullableUDF
 
 object SQLGeometricConstructorFunctions {
